Move run metrics to struct in monitor header
