+++ include/bout/monitor.hxx
@@ -7,6 +7,7 @@
 
 #include <cmath>
 
+class Datafile;
 class Solver;
 
 /// Return true if either \p a is a multiple of \p b or vice-versa
@@ -68,6 +69,57 @@ private:
   bool is_added = false; ///< Set to true when Monitor is added to a Solver
   BoutReal timestep;
   int freq;
+
 };
+
+struct RunMetrics {
+  public:
+  /// cumulative wall clock time in seconds
+  BoutReal t_elapsed = 0;
+  /// time step's wall clock time in seconds
+  BoutReal wtime = 0;
+
+  /// number of RHS calls
+  int ncalls = 0;
+  /// number of RHS calls for fast timescale
+  int ncalls_e = 0;
+  /// number of RHS calls for slow timescale
+  int ncalls_i = 0;
+
+  /// wall time spent calculating RHS
+  BoutReal wtime_rhs = 0;
+  /// wall time spent inverting Laplacian
+  BoutReal wtime_invert = 0;
+  /// wall time spent communicating (part of RHS)
+  BoutReal wtime_comms = 0;
+  /// wall time spent on I/O
+  BoutReal wtime_io = 0;
+
+  // Derived metrics
+
+  /// wall time per RHS evaluation
+  BoutReal wtime_per_rhs = 0;
+  /// wall time per fast timescale RHS evaluation
+  BoutReal wtime_per_rhs_e = 0;
+  /// wall time per slow timescale RHS evaluation
+  BoutReal wtime_per_rhs_i = 0;
+
+  /*!
+   * Adds variables to the output file, for post-processing
+   */
+  void outputVars(Datafile &file);
+
+  /*!
+   * Calculates derived metrics
+   */
+  void calculateDerivedMetrics();
+
+  /*!
+   * Write job progress to screen
+   */
+  void writeProgress(BoutReal simtime, bool output_split);
+
+};
+
 
 #endif // __MONITOR_H__
